Add setting for mp3 quality, add half working ff/rew
git-svn-id: http://svn.mythtv.org/svn/trunk@15 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/RingBuffer.cpp
@@ -148,7 +148,7 @@ int RingBuffer::Write(const void *buf, int count)
     return ret;
 }
 
-long long RingBuffer::Seek(long pos, int whence)
+long long RingBuffer::Seek(long long pos, int whence)
 {
     long long ret = -1;
     if (normalfile)
